+++ python/dremio/flight/query.py
@@ -16,7 +16,7 @@
 import logging
 from pyarrow import flight
 from pandas import DataFrame, concat
-from dremio.flight.dremio_flight_connection import DremioFlightEndpointConnection
+from dremio.flight.connection import DremioFlightEndpointConnection
 
 logging.basicConfig(level=logging.INFO)
