+++ python/dremio/middleware/auth.py
