+++ python/dremio/middleware/cookie.py
